Make React Context work
The previous commit was made due to the implementation of React Context
but it wasn't completely implemented. In this commit the app is fully
functional, Context is correctly applied and Lifting State is implemented
to change the only useState hook of App.js from the Footer.js component.

The components of the UI have been modularized and each one is in a separate
file.