Fix bug displaying a wrong date
Every time a card was clicked, the date associated to that card
wasn't displayed in the Main component. So I added a global date
state (that resides in the App file) which is used in the function
changeDate(). This function is passed as a prop through Aside and
AsideGroup just to reach the AsideCard component, where is implemented
whenever the card is clicked.